fix IPoE over v4 for erouter0 is failing and triggering dnsmasq restart ( OFW-5502 )
[Issue]
IPOE over v4 is failing

[Description]
IPOE is  retrieving the mac address to send the echo packet from the arp cache.
But in some case the apr table is not updated with the v4 address of the default gateway.
The arp table was not updated as the ping was initiated with the erouter ip address instead of BNG ip.

[Solution]
Corrected the ip to which ping is initiated and the arp table updation should happen before the start of echo packet.

[Testing]
1) Mv3 Gateway is enabled with IPoE with release build
2) Upgrade the device .
3) Monitor IHCLog.txt.0 and dnsmasq pid after firmware upgrade

[Upstream Candidate]
Yes

Signed-off-by: Sreeja S Kumar (srekumar.contractor@libertyglobal.com)